move app delegate to expected location
