+++ ios/Artsy/AppDelegate.h
