+++ ios/Artsy/Networking/API_Modules/ARUserManager.m
@@ -16,7 +16,7 @@
 #import "ARKeychainable.h"
 #import "ARSystemTime.h"
 #import "ARLogger.h"
-#import "ARAppDelegate.h"
+#import "AppDelegate.h"
 
 #import "MTLModel+JSON.h"
 #import "AFHTTPRequestOperation+JSON.h"
